interviewx/ai-models/text-analysis/requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Google Generative AI (Gemini)
google-generativeai==0.3.2

# Natural Language Processing
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2

# Text Analysis and Statistics
textstat==0.7.3
scikit-learn==1.3.0
numpy==1.24.3
scipy==1.11.2

# Language Detection
langdetect==1.0.9

# Additional NLP Tools
wordcloud==1.9.2
gensim==4.3.2

# Text Processing
regex==2023.8.8
unidecode==1.3.6

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development and Testing
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Performance monitoring
psutil==5.9.5
memory-profiler==0.61.0

# Logging
colorlog==6.7.0

# Data handling
pandas==2.1.1
openpyxl==3.1.2

# spaCy English model (install separately)
# python -m spacy download en_core_web_sm

# Optional: Advanced language models (uncomment if needed)
# transformers==4.34.0
# torch==2.0.1

# Optional: Grammar checking (uncomment if using external grammar API)
# language-tool-python==2.7.1

# Optional: Additional text analysis
# pyenchant==3.2.2
# pattern==3.6

# NLTK data downloads (handled in code)
# punkt, stopwords, wordnet, averaged_perceptron_tagger, etc.